- MySQL reserved words refactoring. - Test updates.
