+++ liquibase-standard/src/test/groovy/liquibase/database/core/MariaDBDatabaseSpec.groovy
@@ -30,14 +30,14 @@ class MariaDBDatabaseSpec extends Specification {
         null                      || 'COL1'           | Column     || 'col1'           | 'COL1'
         null                      || 'Col1'           | Column     || 'col1'           | 'Col1'
         null                      || 'col with space' | Column     || 'col with space' | '`col with space`'
-        null                      || 'period'         | Column     || 'period'         | '`period`'
-        null                      || 'PERIOD'         | Column     || 'period'         | '`PERIOD`'
+        null                      || 'period'         | Column     || 'period'         | 'period'
+        null                      || 'PERIOD'         | Column     || 'period'         | 'PERIOD'
         QUOTE_ONLY_RESERVED_WORDS || 'col1'           | Column     || 'col1'           | 'col1'
         QUOTE_ONLY_RESERVED_WORDS || 'COL1'           | Column     || 'col1'           | 'COL1'
         QUOTE_ONLY_RESERVED_WORDS || 'Col1'           | Column     || 'col1'           | 'Col1'
         QUOTE_ONLY_RESERVED_WORDS || 'col with space' | Column     || 'col with space' | '`col with space`'
-        QUOTE_ONLY_RESERVED_WORDS || 'period'         | Column     || 'period'         | '`period`'
-        QUOTE_ONLY_RESERVED_WORDS || 'PERIOD'         | Column     || 'period'         | '`PERIOD`'
+        QUOTE_ONLY_RESERVED_WORDS || 'period'         | Column     || 'period'         | 'period'
+        QUOTE_ONLY_RESERVED_WORDS || 'PERIOD'         | Column     || 'period'         | 'PERIOD'
         QUOTE_ALL_OBJECTS         || 'col1'           | Column     || 'col1'           | '`col1`'
         QUOTE_ALL_OBJECTS         || 'COL1'           | Column     || 'col1'           | '`COL1`'
         QUOTE_ALL_OBJECTS         || 'Col1'           | Column     || 'col1'           | '`Col1`'
